+++ tool/build_tests.dart
@@ -1,7 +1,7 @@
 import 'package:build_runner/build_runner.dart';
-import 'package:serializable/phase.dart';
+import 'package:serializable/action.dart';
 
 
 main() async {
-  await build([serializablePhase(const ['test/**.dart'])], deleteFilesByDefault: true);
+  await build([serializableAction(const ['test/**.dart'])], deleteFilesByDefault: true);
 }